Get git adapter working, and testing itself
Some thoughts on what's still to do:

- actually use it (it's all duplicate code at the moment)
- create a contract test and an in-memory fake for faster testing of the
  core. How would this work with fetches?!
- move detailed command tests to sit with handlers?
- different init commands for bare / testable repos?
- separate create / open methods on the factory?
:
