+++ packages/adapter-git/src/handlers/handler.ts
@@ -1,0 +1,3 @@
+import { GitRepo } from '../git_repo'
+
+export type Handler<Command, Result = void> = (repo: GitRepo, command: Command) => Promise<Result>
