+++ packages/adapter-git/src/handlers/index.ts
@@ -1,0 +1,6 @@
+export * from './handleInit'
+export * from './handleSetOrigin'
+export * from './handleFetch'
+export * from './handleCommit'
+export * from './handleGetRevision'
+export * from './handleEnsureBranchExists'
