+++ packages/adapter-git/src/index.spec.ts
@@ -1,23 +1,22 @@
 import childProcess from 'child_process'
-import { Init, SetOrigin, Commit, GetRevision, Fetch } from 'git-en-boite-core-port-git'
+import { Commit, Fetch, GetRevision, Init, SetOrigin } from 'git-en-boite-core-port-git'
 import {
-  assertThat,
   fulfilled,
   hasProperty,
+  isRejectedWith,
   matchesPattern,
   promiseThat,
-  rejected,
   startsWith,
-  isRejectedWith,
 } from 'hamjest'
 import path from 'path'
 import { promisify } from 'util'
 
-import { LocalGitRepo } from './local_git_repo'
+import { GitRepo } from './git_repo'
+import { GitRepoFactory, TestableGitRepoFactory } from './index'
 
 const exec = promisify(childProcess.exec)
-describe(LocalGitRepo.name, () => {
-  const root = path.resolve(__dirname, '../../tmp')
+describe(GitRepo.name, () => {
+  const root = path.resolve(__dirname, '../tmp')
 
   beforeEach(async () => {
     await exec(`rm -rf ${root}`)
@@ -27,7 +26,7 @@ describe(LocalGitRepo.name, () => {
     describe(Init.name, () => {
       it('creates a new bare repo with conservative garbage collection settings', async () => {
         const repoPath = path.resolve(root, 'a-repo-id')
-        const git = await LocalGitRepo.openForCommands(repoPath)
+        const git = await new GitRepoFactory().open(repoPath)
         await git(Init.bareRepo())
         await promiseThat(
           exec('git config --get core.bare', { cwd: repoPath }),
@@ -47,7 +46,7 @@ describe(LocalGitRepo.name, () => {
     describe(SetOrigin.name, () => {
       it('creates a remote called origin pointing to the URL', async () => {
         const repoPath = path.resolve(root, 'a-repo-id')
-        const git = await LocalGitRepo.openForCommands(repoPath)
+        const git = await new GitRepoFactory().open(repoPath)
         const repoUrl = 'git@host/repo'
         await git(Init.bareRepo())
         await git(SetOrigin.toUrl(repoUrl))
@@ -65,7 +64,7 @@ describe(LocalGitRepo.name, () => {
 
         beforeEach(async () => {
           originUrl = path.resolve(root, 'remote', 'a-repo-id')
-          const origin = await LocalGitRepo.openForCommands(originUrl)
+          const origin = await new TestableGitRepoFactory().open(originUrl)
           await origin(Init.normalRepo())
           await origin(Commit.withAnyMessage())
           latestCommit = await origin(GetRevision.forCurrentBranch())
@@ -73,7 +72,7 @@ describe(LocalGitRepo.name, () => {
 
         it('fetches commits from the origin remote', async () => {
           const repoPath = path.resolve(root, 'a-repo-id')
-          const git = await LocalGitRepo.openForCommands(repoPath)
+          const git = await new GitRepoFactory().open(repoPath)
           await git(Init.bareRepo())
           await git(SetOrigin.toUrl(originUrl))
           await git(Fetch.fromOrigin())
@@ -85,7 +84,7 @@ describe(LocalGitRepo.name, () => {
 
         it('fails when the remote does not exist', async () => {
           const repoPath = path.resolve(root, 'a-repo-id')
-          const git = await LocalGitRepo.openForCommands(repoPath)
+          const git = await new GitRepoFactory().open(repoPath)
           await git(Init.bareRepo())
           await git(SetOrigin.toUrl('invalid-remote-url'))
           await promiseThat(
@@ -96,26 +95,6 @@ describe(LocalGitRepo.name, () => {
           )
         })
       })
-    })
-  })
-
-  describe('running arbitrary git commands', () => {
-    it('returns a promise of the result', async () => {
-      const repoId = 'a-new-repo-id'
-      const repoPath = path.resolve(root, repoId)
-      const repo = await LocalGitRepo.open(repoPath)
-      const result = await repo.execGit('init')
-      assertThat(result.stdout, startsWith('Initialized empty Git repository'))
-    })
-
-    it('raises any error', async () => {
-      const repoId = 'a-repo-id'
-      const repoPath = path.resolve(root, repoId)
-      const repo = await LocalGitRepo.open(repoPath)
-      await promiseThat(
-        repo.execGit('not-a-command'),
-        rejected(hasProperty('message', matchesPattern('is not a git command'))),
-      )
     })
   })
 })
